EM pro 50 advance

+++ src/main/java/it/usna/shellyscan/model/device/g2/modules/EM1Manager.java
@@ -120,7 +120,7 @@ public class EM1Manager {
 	
 	//l.add("(BTHomeSensor.GetConfig [" + s.getId() + "-" + s.getObjId() + "])/rpc/BTHomeSensor.GetCon)fig?id=" + s.getId());
 
-	public record TimedData(long timestamp, float ... value) {
+	public record TimedData(int timestamp, float ... value) {
 		@Override
 		public String toString() {
 			return timestamp + "-" + value[0];
